restructured the loginModal and map components to match folder structure and deleted dummy files
