--- src/components/atoms/Button_dummy.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const button_dummy = () => {
-  return <div></div>;
-};
-
-export default button_dummy;
