--- src/components/atoms/Input_dummy.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const input_dummy = () => {
-  return <div></div>;
-};
-
-export default input_dummy;
