+++ src/components/atoms/loginButton.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useState } from 'react';
 import LoginModal from '../molecules/loginModal';
