--- src/components/molecules/SearchInput_dummy.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const searchInput_dummy = () => {
-  return <div></div>;
-};
-
-export default searchInput_dummy;
