+++ src/components/molecules/loginModal.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { useUser } from "../../../contexts/UserContext"
+import { useUser } from "../../contexts/UserContext"
 
 function LoginModal({ isOpen, handleClose }) {
