--- src/components/organisms/Header_dummy.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Header_dummy = () => {
-  return <div></div>;
-};
-
-export default Header_dummy;
