--- src/contexts/Context_dummy.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const context_dummy = () => {
-  return <div></div>;
-};
-
-export default context_dummy;
